+++ .eslintrc.js
@@ -4,6 +4,7 @@ module.exports = {
     jest: true,
   },
   rules: {
+    'linebreak-style': 0,
     'import/no-unresolved': 0,
     'import/extensions': 0,
   },