+++ packages/react-static-plugin-source-filesystem/src/node.api.js
@@ -14,14 +14,23 @@ export default ({
     const { config, stage, debug } = state
     if (!location) {
       throw new Error(
-        'react-static-plugin-source-filesystem: A valid `location` directory is required to use this plugin'
+        'react-static-plugin-source-filesystem: A valid `location` directory ' +
+          'is required to use this plugin'
       )
     }
     // Make a glob extension to get all pages with the set extensions from the pages directory
+    // It should be a directory, not index.js inside that directory. This will
+    // happen when using .resolve in some instances
+    if (/index\.js$/.test(location)) {
+      location = nodePath.dirname(location)
+    }
+
+    // Make a glob extension to get all pages with the set extensions from the
+    // pages directory
     const globExtensions = [...config.extensions, ...extensions]
       .map(ext => `${ext.slice(1)}`) // cut off the period of the extension
       .join(',') // join them for the glob string
-    const pagesGlob = `${location}/**/*.{${globExtensions}}`
+    const pagesGlob = nodePath.join(location, '**', `*.{${globExtensions}}`)
     // Get the pages
 
     if (debug) {
@@ -37,12 +46,20 @@ export default ({
           const originalPath = page
           // Glob path will always have unix style path, convert to windows if necessary
           const template = nodePath.resolve(page)
-          // Make sure the path is relative to the root of the site
-          let path = page.replace(`${location}`, '').replace(/\..*/, '')
-          // turn windows paths back to unix
-          path = path.split('\\').join('/')
+          // Make sure the path is relative to the location root
+          let path = nodePath.relative(location, template)
+          // Cutoff the extension
+          path = nodePath.join(
+            nodePath.dirname(path),
+            nodePath.basename(path, nodePath.extname(path))
+          )
+          // Ensure paths are unix
+          path = path.split(nodePath.sep).join(nodePath.posix.sep)
+          // Make sure it starts with a slash
+          path = path[0] === '/' ? path : `/${path}`
           // Turn `/index` paths into roots`
           path = path.replace(/\/index$/, '/')
+          // Add the path prefix
           path = pathPrefix ? pathJoin(pathPrefix, path) : path
           // Return the route
           return createRoute({