+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Improved
 
 - Fixed issues with custom config specified using `-c` or `-config` is not loaded
+- `react-static-plugin-source-filesystem`: Fixed resolution on Windows
+- `react-static`: Fixed building chunks on Windows
+- `react-static`: Fixed production build on Windows
 
 # 7.0.7